+++ src/datasources/asset/data-sources/list-assets/ListAssetsDataSource.ts
@@ -2,11 +2,13 @@ import { DataQueryRequest, DataFrameDTO, DataSourceInstanceSettings } from '@gra
 import { BackendSrv, getBackendSrv, getTemplateSrv, TemplateSrv } from '@grafana/runtime';
 import { AssetDataSourceBase } from '../AssetDataSourceBase';
 import { AssetDataSourceOptions, AssetQuery, AssetQueryType } from '../../types/types';
-import { ListAssetsQuery } from '../../types/ListAssets.types';
+import { ListAssetsQuery, OutputType, QueryListAssetRequestBody } from '../../types/ListAssets.types';
 import { AssetModel, AssetsResponse } from '../../../asset-common/types';
 import { getWorkspaceName } from '../../../../core/utils';
 import { transformComputedFieldsQuery } from '../../../../core/query-builder.utils';
 import { QUERY_LIMIT } from 'datasources/asset/constants/constants';
+import { defaultListAssetsQuery } from 'datasources/asset/defaults';
+import { TAKE_LIMIT } from 'datasources/asset/constants/ListAssets.constants';
 
 export class ListAssetsDataSource extends AssetDataSourceBase {
   private dependenciesLoadedPromise: Promise<void>;
@@ -24,13 +26,14 @@ export class ListAssetsDataSource extends AssetDataSourceBase {
 
   defaultQuery = {
     type: AssetQueryType.ListAssets,
-    filter: ''
+    filter: '',
+    take: 1000,
   };
 
   async runQuery(query: AssetQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
-    const listAssetsQuery = query as ListAssetsQuery;
+    const listAssetsQuery = this.patchListAssetQuery(query);
     await this.dependenciesLoadedPromise;
-  
+
     if (listAssetsQuery.filter) {
       listAssetsQuery.filter = transformComputedFieldsQuery(
         this.templateSrv.replace(listAssetsQuery.filter, options.scopedVars),
@@ -39,7 +42,19 @@ export class ListAssetsDataSource extends AssetDataSourceBase {
       );
     }
 
-    return this.processListAssetsQuery(listAssetsQuery);
+    if (listAssetsQuery.outputType === OutputType.TotalCount) {
+      return this.processTotalCountAssetsQuery(listAssetsQuery);
+    };
+
+    if (listAssetsQuery.outputType === OutputType.Properties && this.isTakeValid(listAssetsQuery)) {
+      return this.processListAssetsQuery(listAssetsQuery);
+    }
+
+    return {
+      refId: query.refId,
+      name: query.refId,
+      fields: [],
+    };
   }
 
   shouldRunQuery(query: AssetQuery): boolean {
@@ -48,7 +63,8 @@ export class ListAssetsDataSource extends AssetDataSourceBase {
 
   async processListAssetsQuery(query: ListAssetsQuery) {
     const result: DataFrameDTO = { refId: query.refId, fields: [] };
-    const assets: AssetModel[] = await this.queryAssets(query.filter, QUERY_LIMIT);
+    const assetsResponse: AssetsResponse = await this.queryAssets(query.filter, query.take, false);
+    const assets = assetsResponse.assets;
     const workspaces = this.getCachedWorkspaces();
     result.fields = [
       { name: 'id', values: assets.map(a => a.id) },
@@ -65,10 +81,10 @@ export class ListAssetsDataSource extends AssetDataSourceBase {
       { name: 'calibration status', values: assets.map(a => a.calibrationStatus) },
       { name: 'is system controller', values: assets.map(a => a.isSystemController) },
       { name: 'last updated timestamp', values: assets.map(a => a.lastUpdatedTimestamp) },
-      { name: 'location', values: assets.map(a => this.getLocationFromAsset(a))},
+      { name: 'location', values: assets.map(a => this.getLocationFromAsset(a)) },
       { name: 'minionId', values: assets.map(a => a.location.minionId) },
       { name: 'parent name', values: assets.map(a => a.location.parent) },
-      { name: 'workspace', values: assets.map( a => getWorkspaceName( workspaces, a.workspace ) ) },
+      { name: 'workspace', values: assets.map(a => getWorkspaceName(workspaces, a.workspace)) },
       { name: 'supports self calibration', values: assets.map(a => a.supportsSelfCalibration) },
       { name: 'supports external calibration', values: assets.map(a => a.supportsExternalCalibration) },
       { name: 'visa resource name', values: assets.map(a => a.visaResourceName) },
@@ -84,22 +100,34 @@ export class ListAssetsDataSource extends AssetDataSourceBase {
     return result;
   }
 
-  async queryAssets(filter = '', take = -1): Promise<AssetModel[]> {
-    let data = { filter, take };
+  async processTotalCountAssetsQuery(query: ListAssetsQuery) {
+    const response: AssetsResponse = await this.queryAssets(query.filter, QUERY_LIMIT, true);
+    const result: DataFrameDTO = { refId: query.refId, fields: [{ name: "Total count", values: [response.totalCount] }] };
+    return result;
+  }
+
+  async queryAssets(filter = '', take = -1, returnCount = false): Promise<AssetsResponse> {
+    let data: QueryListAssetRequestBody = { filter, take, returnCount };
     try {
-      let response = await this.post<AssetsResponse>(this.baseUrl + '/query-assets', data);
-      return response.assets;
+      const response = await this.post<AssetsResponse>(this.baseUrl + '/query-assets', data);
+      return response;
     } catch (error) {
       throw new Error(`An error occurred while querying assets: ${error}`);
     }
   }
 
-  private getLocationFromAsset(asset: AssetModel): string
-  {
-    if (asset.location.physicalLocation)
-    {
+  private getLocationFromAsset(asset: AssetModel): string {
+    if (asset.location.physicalLocation) {
       return asset.location.physicalLocation;
     }
     return this.systemAliasCache.get(asset.location.minionId)?.alias || '';
+  }
+
+  public patchListAssetQuery(query: AssetQuery): ListAssetsQuery {
+    return { ...defaultListAssetsQuery, ...query } as ListAssetsQuery;
+  }
+
+  private isTakeValid(query: ListAssetsQuery): boolean {
+    return query.take !== undefined && query.take >= 0 && query.take <= TAKE_LIMIT;
   }
 }